# public/openapi.yaml
openapi: 3.0.3
info:
  title: HireU / Talent Platform API
  description: REST API for the hiring platform — jobs, applications, profile, and company dashboard.
  version: 1.0.0
servers:
  - url: /
    description: Current origin (relative)
...
paths:
...
  /api/company/jobs/{id}:
    get:
      tags: [Company]
      summary: Get job by ID (recruiter, own job only)
      security: [{ session: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Job with application count
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Company]
      summary: Update job (recruiter, own job only)
      security: [{ session: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                location: { type: string }
                description: { type: string }
                employmentType: { type: string }
                category: { type: string }
                skills: { type: array, items: { type: string } }
                salary: { type: string }
                currency: { type: string }
                status: { type: string }
      responses:
        '200':
          description: Updated job
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Company]
      summary: Delete job (recruiter, own job only)
      security: [{ session: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Job and its applications deleted
          content:
            application/json:
              schema: { type: object, properties: { message: { type: string } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ----- Applications (candidate) -----
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie (sign in via the app)
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
      headers: {}
    Forbidden:
      description: Authenticated but not allowed (e.g. wrong role)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
    Job:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        company: { type: string }
        location: { type: string }
        description: { type: string }
        employmentType: { type: string }
        category: { type: string }
        skills: { type: array, items: { type: string } }
        salary: { type: string }
        currency: { type: string }
        status: { type: string }
        experienceMin: { type: integer }
        experienceMax: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...